feat: don't walk entire project
