+++ test/nuts/local/pkgDirMatching.nut.ts
@@ -9,7 +9,7 @@ import { TestSession } from '@salesforce/cli-plugins-testkit';
 import * as fs from 'graceful-fs';
 import { expect } from 'chai';
 import { RegistryAccess } from '@salesforce/source-deploy-retrieve';
-import { ShadowRepo } from '../../../src/shared/localShadowRepo';
+import { ShadowRepo } from '../../../src/shared/local/localShadowRepo';
 
 describe('verifies exact match of pkgDirs', () => {
   const registry = new RegistryAccess();
