+++ test/unit/localDetectMovedFiles.test.ts
@@ -12,7 +12,7 @@ import git from 'isomorphic-git';
 import { expect } from 'chai';
 import sinon = require('sinon');
 import { RegistryAccess } from '@salesforce/source-deploy-retrieve';
-import { ShadowRepo } from '../../src/shared/localShadowRepo';
+import { ShadowRepo } from '../../src/shared/local/localShadowRepo';
 
 /* eslint-disable no-unused-expressions */
